\name{dawai-internal}
\alias{bootstrapsamples}
\alias{checks}
\alias{classify}
\alias{constrMatrix}
\alias{getmeans}
\alias{getvariances}
\alias{posterior}
\alias{print.restrictions}
\alias{print.rlda}
\alias{print.rqda}
\alias{rest.est}
\alias{restrictions}
\alias{restrictions.default}
\title{Internal dawai functions}
\description{
Internal dawai functions
}
\details{
These are not to be called by the user.
}
\keyword{ internal }