\name{dawai-package}
\alias{dawai-package}
\alias{dawai}
\docType{package}
\title{
Discriminant analysis with additional information
}
\description{
This package performs linear and quadratic discriminant analysis with additional information expressed as inequality constraints among the populations means and computes several estimations of the true error rate
}
\details{
Package: dawai

Type: Package

Version: 1.2.3

Date: 2018-01-16

License: GPL-2 | GPL-3


For a complete list of functions with individual help pages, use \code{library(help = "dawai")}.
}
\author{
David Conde, Miguel A. Fernandez, Bonifacio Salvador

Maintainer: David Conde <dconde@eio.uva.es>
}
\references{
Conde, D., Fernandez, M. A., Rueda, C., and Salvador, B. (2012). Classification of samples into two or more ordered populations with application to a cancer trial. \emph{Statistics in Medicine}, \bold{31}, 3773-3786.

Conde, D., Fernandez, M. A., Salvador, B., and Rueda, C. (2015). dawai: An R Package for Discriminant Analysis with Additional Information. \emph{Journal of Statistical Software}, \bold{66(10)}, 1-19. URL http://www.jstatsoft.org/v66/i10/.

Conde, D., Salvador, B., Rueda, C. , and Fernandez, M. A. (2013). Performance and estimation of the true error rate of classification rules built with additional information. An application to a cancer trial. \emph{Statistical Applications in Genetics and Molecular Biology}, \bold{12(5)}, 583-602.

Fernandez, M. A., Rueda, C., Salvador, B. (2006). Incorporating additional information to normal linear discriminant rules. \emph{Journal of the American Statistical Association}, \bold{101}, 569-577.
}
\keyword{ package }
\keyword{ multivariate }
\keyword{ classif }
